+++ ryan-scripts/TUFLOW-python/LogSummary.py
@@ -12,8 +12,8 @@ import gc
 from pathlib import Path
 import os
 
-from ryan_library.scripts.tuflow.tuflow_logsummary import main_processing
-from ryan_library.scripts.wrapper_utils import (
+from ryan_library.orchestrators.tuflow.tuflow_logsummary import main_processing
+from ryan_library.functions.wrapper_utils import (
     CommonWrapperOptions,
     add_common_cli_arguments,
     change_working_directory,
