+++ ryan-scripts/TUFLOW-python/POMM-mean-max-aep-dur.py
@@ -29,8 +29,8 @@ import argparse
 import gc
 import os
 
-from ryan_library.scripts.tuflow.pomm_max_items import export_mean_peak_report
-from ryan_library.scripts.wrapper_utils import (
+from ryan_library.orchestrators.tuflow.pomm_max_items import export_mean_peak_report
+from ryan_library.functions.wrapper_utils import (
     CommonWrapperOptions,
     PommPeakWrapperDefaults,
     add_common_cli_arguments,
