+++ ryan-scripts/TUFLOW-python/POMM-med-max-aep-dur.py
@@ -30,8 +30,8 @@ import argparse
 import gc
 import os
 
-from ryan_library.scripts.tuflow.pomm_max_items import export_median_peak_report
-from ryan_library.scripts.wrapper_utils import (
+from ryan_library.orchestrators.tuflow.pomm_max_items import export_median_peak_report
+from ryan_library.functions.wrapper_utils import (
     CommonWrapperOptions,
     PommPeakWrapperDefaults,
     add_common_cli_arguments,
