+++ ryan-scripts/TUFLOW-python/PO_combine.py
@@ -2,7 +2,7 @@
 """
 Wrapper Script: Combine TUFLOW PO Results.
 
-This script acts as a mutable wrapper for `ryan_library.scripts.tuflow.po_combine`.
+This script acts as a mutable wrapper for `ryan_library.orchestrators.tuflow.po_combine`.
 It allows users to hard-code default configurations for merging TUFLOW "PO" (Plot Output) CSV files,
 while still providing command-line overrides for automation/batch processing.
 
@@ -31,8 +31,8 @@ import argparse
 import gc
 import os
 
-from ryan_library.scripts.tuflow.po_combine import main_processing
-from ryan_library.scripts.wrapper_utils import (
+from ryan_library.orchestrators.tuflow.po_combine import main_processing
+from ryan_library.functions.wrapper_utils import (
     CommonWrapperOptions,
     add_common_cli_arguments,
     change_working_directory,
