+++ ryan-scripts/TUFLOW-python/TUFLOW-find-closure-durations.py
@@ -22,8 +22,8 @@ import gc
 import os
 
 
-from ryan_library.scripts.tuflow.closure_durations import run_closure_durations
-from ryan_library.scripts.wrapper_utils import (
+from ryan_library.orchestrators.tuflow.closure_durations import run_closure_durations
+from ryan_library.functions.wrapper_utils import (
     CommonWrapperOptions,
     add_common_cli_arguments,
     change_working_directory,
