+++ ryan-scripts/TUFLOW-python/TUFLOW_Culvert_Timeseries.py
@@ -21,8 +21,8 @@ import argparse
 import gc
 import os
 
-from ryan_library.scripts.tuflow.tuflow_culverts_timeseries import main_processing
-from ryan_library.scripts.wrapper_utils import (
+from ryan_library.orchestrators.tuflow.tuflow_culverts_timeseries import main_processing
+from ryan_library.functions.wrapper_utils import (
     CommonWrapperOptions,
     add_common_cli_arguments,
     change_working_directory,
