+++ ryan_library/__init__.py
@@ -1,3 +1,4 @@
 from .functions import *
+from .orchestrators import *
 from .scripts import *
 from .processors import *
