+++ ryan_library/functions/tuflow/notebook_helpers.py
@@ -11,7 +11,7 @@ from ryan_library.functions.loguru_helpers import setup_logger
 from ryan_library.functions.tuflow.tuflow_common import collect_files, process_files_in_parallel, process_file
 from ryan_library.processors.tuflow.base_processor import BaseProcessor
 from ryan_library.processors.tuflow.processor_collection import ProcessorCollection
-from ryan_library.scripts.tuflow.tuflow_culverts_mean import find_culvert_aep_dur_mean, find_culvert_aep_mean_max
+from ryan_library.orchestrators.tuflow.tuflow_culverts_mean import find_culvert_aep_dur_mean, find_culvert_aep_mean_max
 
 import pandas as pd
 import matplotlib.pyplot as plt
