+++ ryan_library/orchestrators/rorb/closure_durations.py
@@ -1,4 +1,4 @@
-# ryan_library\scripts\RORB\closure_durations.py
+# ryan_library\orchestrators\rorb\closure_durations.py
 
 from datetime import datetime
 from pathlib import Path
