+++ ryan_library/orchestrators/tuflow/tuflow_results_styling.py
@@ -1,4 +1,4 @@
-# ryan_library/scripts/tuflow/tuflow_results_styling.py
+# ryan_library/orchestrators/tuflow/tuflow_results_styling.py
 """
 TUFLOW Results Styling.
 
@@ -43,7 +43,7 @@ class TUFLOWResultsStyler:
 
     def __init__(self, user_qml_overrides: dict[str, str] | None = None) -> None:
         """Initializes the TUFLOWResultsStyler with default styles path and user overrides."""
-        # __file__ resolves to ryan_library/scripts/tuflow/tuflow_results_styling.py
+        # __file__ resolves to ryan_library/orchestrators/tuflow/tuflow_results_styling.py
         # We need the repository root to locate QML files under QGIS-Styles/TUFLOW
         self.default_styles_path: Path = Path(__file__).absolute().parents[3] / "QGIS-Styles" / "TUFLOW"
         logger.debug("Default styles path: {}", self.default_styles_path)
@@ -158,7 +158,7 @@ class TUFLOWResultsStyler:
             )
             styles: list[tuple[str, str]] = cursor.fetchall()
 
-            for style_name, style_qml in styles:
+            for style_name, _style_qml in styles:
                 logger.info(
                     f"Applying style '{style_name}' to layer '{layer_name}'",
                 )
