+++ ryan_library/orchestrators/tuflow/tuflow_timeseries_stability.py
@@ -1,9 +1,8 @@
-# ryan_library/scripts/tuflow/peak_check_po_csvs.py
+# ryan_library/orchestrators/tuflow/tuflow_timeseries_stability.py
 """
-Peak checks for TUFLOW PO timeseries CSVs.
+Timeseries stability checks for TUFLOW outputs.
 
-This module scans PO CSVs, identifies peak timing relative to the end of the series,
-and exports a summary table.
+Currently this supports PO CSV files only; 1D Q support can be added later.
 """
 
 from __future__ import annotations
@@ -15,13 +14,15 @@ from typing import Literal
 
 import pandas as pd
 from loguru import logger
+from pandas import DataFrame
 
 from ryan_library.functions.loguru_helpers import setup_logger
 from ryan_library.functions.misc_functions import ExcelExporter
 from ryan_library.functions.tuflow.po_timeseries_checks import (
-    PeakCheckConfig,
-    analyze_peak_csv,
-    flatten_peak_results,
+    StabilityCheckConfig,
+    StabilityCheckResult,
+    analyze_stability_csv,
+    flatten_stability_results,
 )
 
 
@@ -41,9 +42,9 @@ def _collect_files(paths_to_process: Sequence[Path], csv_glob: str) -> list[Path
     return sorted(files)
 
 
-def _analyze_peak_worker(path_str: str, config: PeakCheckConfig) -> list[dict[str, object]]:
-    results = analyze_peak_csv(path=Path(path_str), config=config)
-    return flatten_peak_results(results=results)
+def _analyze_stability_worker(path_str: str, config: StabilityCheckConfig) -> list[dict[str, object]]:
+    results: list[StabilityCheckResult] = analyze_stability_csv(path=Path(path_str), config=config)
+    return flatten_stability_results(results=results)
 
 
 def main_processing(
@@ -55,9 +56,11 @@ def main_processing(
     location_include: Sequence[str] = (),
     location_exclude: Sequence[str] = (),
     location_case_sensitive: bool = False,
-    warn_2hours: float = 2.0,
-    warn_1hour: float = 1.0,
     flat_tol: float = 1e-6,
+    diff_rel_tol: float = 0.01,
+    diff_abs_tol: float = 1e-6,
+    max_sign_changes: int = 2,
+    min_points: int = 5,
     max_workers: int | None = None,
     chunksize: int = 1,
     console_log_level: str = "INFO",
@@ -65,7 +68,7 @@ def main_processing(
     export_mode: Literal["excel", "parquet", "both"] = "excel",
 ) -> None:
     """
-    Run peak checks for PO timeseries CSV files and export a summary.
+    Run stability checks for PO timeseries CSV files and export a summary.
 
     Args:
         paths_to_process: Directories to scan for PO CSVs.
@@ -75,37 +78,41 @@ def main_processing(
         location_include: Optional location allow-list.
         location_exclude: Optional location block-list.
         location_case_sensitive: Whether location filtering is case-sensitive.
-        warn_2hours: Threshold (hours) for WARN_2H.
-        warn_1hour: Threshold (hours) for WARN_1H.
-        flat_tol: Tolerance for treating peak deviations as flat.
+        flat_tol: Range threshold for considering a series flat.
+        diff_rel_tol: Relative tolerance (of range) for ignoring step noise.
+        diff_abs_tol: Absolute tolerance for ignoring step noise.
+        max_sign_changes: Maximum sign changes allowed before flagging unstable.
+        min_points: Minimum points required for stability evaluation.
         max_workers: Process pool size for parallel analysis.
         chunksize: Task chunk size for the process pool.
         console_log_level: Loguru console log level.
         output_dir: Optional output directory for exports.
         export_mode: "excel", "parquet", or "both".
     """
-    config = PeakCheckConfig(
+    config = StabilityCheckConfig(
         datatype_include=datatype_include,
         datatype_case_sensitive=datatype_case_sensitive,
         location_include=location_include,
         location_exclude=location_exclude,
         location_case_sensitive=location_case_sensitive,
-        warn_2hours=warn_2hours,
-        warn_1hour=warn_1hour,
         flat_tol=flat_tol,
+        diff_rel_tol=diff_rel_tol,
+        diff_abs_tol=diff_abs_tol,
+        max_sign_changes=max_sign_changes,
+        min_points=min_points,
     )
 
     with setup_logger(console_log_level=console_log_level):
-        files = _collect_files(paths_to_process=paths_to_process, csv_glob=csv_glob)
+        files: list[Path] = _collect_files(paths_to_process=paths_to_process, csv_glob=csv_glob)
         if not files:
             logger.info(f"No files matched '{csv_glob}' in the provided directories.")
             return
 
-        logger.info(f"Processing {len(files)} PO CSV file(s) for peak checks.")
+        logger.info(f"Processing {len(files)} PO CSV file(s) for stability checks.")
         all_rows: list[dict[str, object]] = []
         with cf.ProcessPoolExecutor(max_workers=max_workers) as executor:
             for rows in executor.map(
-                _analyze_peak_worker,
+                _analyze_stability_worker,
                 (str(path) for path in files),
                 (config for _ in files),
                 chunksize=chunksize,
@@ -123,16 +130,20 @@ def main_processing(
             "status",
             "datatype",
             "location",
-            "peak_kind",
-            "peak_value",
-            "peak_time",
-            "end_time",
-            "hours_from_end",
+            "points",
+            "sign_changes",
+            "nonzero_steps",
+            "value_range",
+            "delta_tol",
+            "value_min",
+            "value_max",
             "start_value",
             "end_value",
-            "end_minus_start",
-            "peak_above_start",
-            "end_pct_of_peak",
+            "start_time",
+            "end_time",
+            "time_span",
+            "max_abs_step",
+            "mean_abs_step",
             "raw_run_code",
             "trim_run_code",
             "data_type",
@@ -147,14 +158,14 @@ def main_processing(
         ordered: list[str] = [c for c in first_cols if c in out_df.columns] + [
             c for c in out_df.columns if c not in first_cols
         ]
-        out_df = out_df.reindex(columns=ordered)
+        out_df: DataFrame = out_df.reindex(columns=ordered)
 
         ExcelExporter().save_to_excel(
             data_frame=out_df,
-            file_name_prefix="peaks_summary",
+            file_name_prefix="timeseries_stability",
             sheet_name="Summary",
             output_directory=output_dir,
             export_mode=export_mode,
             parquet_compression="gzip",
         )
-        logger.info("Peak summary export complete.")
+        logger.info("Timeseries stability export complete.")
